ENH: make tkagg blocking work
The issue is that we need a tk window to grab the mainloop.  In the upstream
tkagg backend we get a manager from Gcf, but in mpl-gui we are not populating
Gcf so we need to patch the tkagg backend.

To get the run-time patching to work as we also needed to add a path to
directly use the Backend classes in the modules rather than re-constituting
another class!